feat: show quick entry layout edit button in respective modal and remove from settings
